moved from SchemaRegistryRestApplication to RestApp
