+++ kcbq-connector/src/test/java/com/wepay/kafka/connect/bigquery/integration/BigQuerySinkConnectorIT.java
@@ -21,22 +21,18 @@ package com.wepay.kafka.connect.bigquery.integration;
 
 import com.wepay.kafka.connect.bigquery.config.BigQuerySinkConfig;
 import com.wepay.kafka.connect.bigquery.integration.utils.BucketClearer;
+import com.wepay.kafka.connect.bigquery.integration.utils.SchemaRegistryTestUtils;
 import com.wepay.kafka.connect.bigquery.integration.utils.TableClearer;
 import com.wepay.kafka.connect.bigquery.retrieve.IdentitySchemaRetriever;
 import com.wepay.kafka.connect.bigquery.utils.FieldNameSanitizer;
 import io.confluent.connect.avro.AvroConverter;
 import io.confluent.kafka.formatter.AvroMessageReader;
-import io.confluent.kafka.schemaregistry.ClusterTestHarness;
-import io.confluent.kafka.schemaregistry.CompatibilityLevel;
-import io.confluent.kafka.schemaregistry.RestApp;
-import io.confluent.kafka.schemaregistry.rest.SchemaRegistryConfig;
 import kafka.common.MessageReader;
 import org.apache.kafka.clients.producer.KafkaProducer;
 import org.apache.kafka.clients.producer.Producer;
 import org.apache.kafka.clients.producer.ProducerConfig;
 import org.apache.kafka.clients.producer.ProducerRecord;
 import org.apache.kafka.common.serialization.Serdes;
-import org.apache.kafka.common.utils.Utils;
 import org.apache.kafka.connect.runtime.ConnectorConfig;
 import org.apache.kafka.connect.runtime.SinkConnectorConfig;
 import org.apache.kafka.test.IntegrationTest;
@@ -162,7 +158,7 @@ public class BigQuerySinkConnectorIT {
 
   // Share a single embedded Connect and Schema Registry cluster for all test cases to keep the runtime down
   private static BaseConnectorIT testBase;
-  private static EmbeddedSchemaRegistry schemaRegistry;
+  private static SchemaRegistryTestUtils schemaRegistry;
   private static String schemaRegistryUrl;
 
   private final String testCase;
@@ -188,7 +184,7 @@ public class BigQuerySinkConnectorIT {
     testBase = new BaseConnectorIT() {};
     testBase.startConnect();
 
-    schemaRegistry = new EmbeddedSchemaRegistry(testBase.connect.kafka().bootstrapServers());
+    schemaRegistry = new SchemaRegistryTestUtils(testBase.connect.kafka().bootstrapServers());
 
     schemaRegistry.start();
 
@@ -221,7 +217,7 @@ public class BigQuerySinkConnectorIT {
   }
 
   @AfterClass
-  public static void globalCleanup() {
+  public static void globalCleanup() throws Exception {
     if (schemaRegistry != null) {
      schemaRegistry.stop();
     }
